# packages/api-client/openapi.snapshot.json
{
  "components": {
...
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "",
    "title": "DS Platform API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/admin/projects/{id}": {
      "get": {
        "operationId": "ProjectsAdminController_detail",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["ProjectsAdmin"]
      },
      "patch": {
        "operationId": "ProjectsAdminController_update",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["ProjectsAdmin"]
      }
    },
    "/v1/admin/projects/{id}/publish": {
      "post": {
        "operationId": "ProjectsAdminController_publish",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["ProjectsAdmin"]
      }
    },
...
  },
  "servers": [],
...
}